Scope `QState` inside `Queue`
It's been renamed now to just `State`.
